create modcart.php
this is what i have for now, i'm exhausted so I'm going to be done for
the day. Supposedly it adds a new item to the cart if the user has added
one, it then will display the list of items in users cart if a user is
logged in.